% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNM_class_methods.R
\name{check_convergence,MNM-method}
\alias{check_convergence,MNM-method}
\title{Convergence Check for "MNM" Class}
\usage{
\S4method{check_convergence}{MNM}(object)
}
\arguments{
\item{object}{An object of class \code{"MNM"}.}
}
\value{
A numeric value indicating the number of converged parameters.
}
\description{
Checks the convergence status of an object of class \code{"MNM"} based on convergence diagnostics (e.g., Rhat values).
Returns the number of parameters that meet the convergence criterion (Rhat < 1.1).
}
