% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNM_class_methods.R
\name{density,MNM-method}
\alias{density,MNM-method}
\title{Density Plot Method for "MNM" Class}
\usage{
\S4method{density}{MNM}(x, param, ...)
}
\arguments{
\item{x}{An object of class \code{"MNM"}.}

\item{param}{The name of the parameter to plot (e.g., \code{"N[8, 1]"}).}

\item{...}{Additional arguments (not used).}
}
\value{
No return value; displays the density plot for the specified parameter.
}
\description{
Generates the density plot for a specified parameter stored in the MNM object.
}
\examples{
# Calling the density function
# density(y, "N[10, 1]")

}
