% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNM_class_methods.R
\name{logLik,MNM-method}
\alias{logLik,MNM-method}
\title{Log-Likelihood Method for "MNM" Class}
\usage{
\S4method{logLik}{MNM}(object)
}
\arguments{
\item{object}{An object of class \code{"MNM"}.}
}
\value{
A numeric value representing the log-likelihood of the fitted model.
}
\description{
Retrieves the log-likelihood value from an object of class \code{"MNM"}.
}
