% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNM_class_methods.R
\name{predictY}
\alias{predictY}
\alias{predictY,MNM-method}
\title{Predict Fitted Values for "MNM" Class}
\usage{
predictY(object)

\S4method{predictY}{MNM}(object)
}
\arguments{
\item{object}{An object of class \code{"MNM"}.}
}
\value{
An array containing the predicted (fitted) values from the model.
}
\description{
Returns the predicted (fitted) values, denoted as \eqn{Y}, from an object of class \code{"MNM"}.
}
\section{Methods (by class)}{
\itemize{
\item \code{predictY(MNM)}: Predict fitted values for the "MNM" class.

}}
