% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNM_class_methods.R
\name{tracePlot}
\alias{tracePlot}
\alias{tracePlot,MNM-method}
\title{Trace Plot Generic and Method for "MNM" Class}
\usage{
tracePlot(x, param, ...)

\S4method{tracePlot}{MNM}(x, param, ...)
}
\arguments{
\item{x}{An object of class \code{"MNM"}.}

\item{param}{The name of the parameter to plot (e.g., \code{"N[8, 1]"}).}

\item{...}{Additional arguments (not used for \code{"MNM"}).}
}
\value{
A trace plot for the specified parameter is displayed.
}
\description{
The \code{tracePlot} function generates trace plots for parameters in objects. This
documentation includes the generic and the method for the \code{"MNM"} class.
}
\section{Methods (by class)}{
\itemize{
\item \code{tracePlot(MNM)}: Method for objects of class \code{"MNM"}.

}}
\examples{
\dontrun{
# Assuming `y` is an object of class "MNM" with plots stored
tracePlot(y, "N[8, 1]")  # Generates a trace plot for parameter N[8, 1]
}
}
